requirements.txt
alembic==1.0.3
aniso8601==4.0.1
argcomplete==1.9.3
boto3==1.9.51
botocore==1.12.51
certifi==2018.10.15
cfn-flip==1.1.0.post1
chardet==3.0.4
Click==7.0
docutils==0.14
durationpy==0.5
Flask==1.0.2
flask-blueprint==1.2.6
Flask-Cors==3.0.7
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.3.2
future==0.16.0
hjson==3.0.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.16.3
marshmallow-sqlalchemy==0.15.0
mimerender==0.6.0
placebo==0.8.2
psycopg2==2.7.6.1
python-dateutil==2.6.1
python-dotenv==0.9.1
python-editor==1.0.3
python-mimeparse==1.6.0
python-slugify==1.2.4
pytz==2018.7
PyYAML==3.13
requests==2.20.1
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.14
SQLAlchemy-Utils==0.33.8
toml==0.10.0
tqdm==4.19.1
troposphere==2.3.4
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
zappa==0.47.1
